requirements.txt
flask>=3.0
requests>=2.28
python-dotenv>=1.0
